requirements.txt
openai
streamlit_drawable_canvas
pandas
numpy
Pillow
PyPDF2
extcolors
pytest
pylint
ipdb
google-cloud-storage
pytz
python-multipart
opencv-python-headless
shapely==2.0.1
pystac-client==0.6.
rasterio==1.3.5
rioxarray==0.13.3




